[coreAudio/audioStreamDCA/Eac3/MP3/PCM] Cast uint64_t to double instead of float for file offset and dts calculation, empty buffer on seek for E-AC3 and DCA too

+++ avidemux_core/ADM_coreAudio/src/ADM_audioStreamMP3.cpp
@@ -31,12 +31,12 @@ ADM_audioStreamMP3::ADM_audioStreamMP3(WAVHeader *header,ADM_audioAccess *access
     if(access->isCBR()==true && access->canSeekOffset()==true)
     {
         // We can compute the duration from the length
-        float size=access->getLength();
-              size/=header->byterate; // Result is in second
-              size*=1000;
-              size*=1000; // s->us
-              durationInUs=(uint64_t)size;
-              return;
+        double size=access->getLength();
+        size/=header->byterate; // Result is in second
+        size*=1000;
+        size*=1000; // s->us
+        durationInUs=(uint64_t)size;
+        return;
     }
     // and built vbr map if needed
     // The 2 conditions below means there is no gap i.e. avi style stream
