+++ avidemux_core/ADM_coreAudio/src/ADM_audioStreamPCM.cpp
@@ -15,11 +15,11 @@ ADM_audioStreamPCM::ADM_audioStreamPCM(WAVHeader *header,ADM_audioAccess *access
     if(access->canGetDuration()==false)
     {
         // We can compute the duration from the length
-        float size=access->getLength();
-              size/=header->byterate; // Result is in second
-              size*=1000;
-              size*=1000; // s->us
-              durationInUs=(uint64_t)size;
+        double size=access->getLength();
+        size/=header->byterate; // Result is in second
+        size*=1000;
+        size*=1000; // s->us
+        durationInUs=(uint64_t)size;
     }
 }
